fix(job): address PR review comments for GKE NAP job submit
