+++ pkg/config/hardware.go
@@ -570,78 +570,3 @@ func Is3DTorusTPU(machineType string) bool {
 	}
 	return false
 }
-
-// GetTPULimitKey resolves a machine type to its corresponding TPU limit key.
-func GetTPULimitKey(machineType string) string {
-	m := strings.ToLower(machineType)
-
-	for _, k := range sortedTPUMachineFamilies {
-		if strings.Contains(m, k) {
-			return parsedMappings.MachineFamilyToLabelMap[k]
-		}
-	}
-	return "google.com/tpu"
-}
-
-// GetGPULimitKey resolves a machine type and accelerator label to its corresponding GPU limit key.
-func GetGPULimitKey(machineType string, accelLabel string) (string, error) {
-	m := strings.ToLower(machineType)
-	accel := strings.ToLower(accelLabel)
-
-	// Reject TPU machine types statically
-	if IsTPU(machineType) {
-		return "", fmt.Errorf("machine type %q is a TPU machine type, not a GPU machine type", machineType)
-	}
-
-	// Reject known CPU-only machine families if no accelerator is attached
-	if accel == "" {
-		for _, family := range parsedMappings.CPUMachineFamilies {
-			if strings.HasPrefix(m, family) {
-				return "", fmt.Errorf("machine type %q is a CPU-only machine type", machineType)
-			}
-		}
-	}
-
-	// Resolve by machine type family
-	for _, k := range sortedGPUMachineFamilies {
-		if strings.Contains(m, k) {
-			return parsedMappings.MachineFamilyToLabelMap[k], nil
-		}
-	}
-
-	// Fallback based on accelerator label
-	if key, ok := resolveGPULabelFallback(accel); ok {
-		return key, nil
-	}
-
-	if accel != "" {
-		return "", fmt.Errorf("unknown accelerator label: %q", accelLabel)
-	}
-
-	return "nvidia.com/gpu", nil
-}
-
-func resolveGPULabelFallback(accel string) (string, bool) {
-	if strings.Contains(accel, "l4") {
-		return "nvidia-l4", true
-	}
-	if strings.Contains(accel, "a100") {
-		return "nvidia-tesla-a100", true
-	}
-	if strings.Contains(accel, "h100") {
-		return "nvidia-h100-80gb", true
-	}
-	if strings.Contains(accel, "h200") {
-		return "nvidia-h200-141gb", true
-	}
-	if strings.Contains(accel, "b200") {
-		return "nvidia-b200", true
-	}
-	if strings.Contains(accel, "t4") {
-		return "nvidia-tesla-t4", true
-	}
-	if strings.Contains(accel, "v100") {
-		return "nvidia-tesla-v100", true
-	}
-	return "", false
-}
