+++ pkg/config/hardware_test.go
@@ -633,178 +633,3 @@ func TestIs3DTorusTPU(t *testing.T) {
 		})
 	}
 }
-
-func TestGetTPULimitKey(t *testing.T) {
-	tests := []struct {
-		machineType string
-		wantKey     string
-	}{
-		{
-			machineType: "ct6e-standard-4t",
-			wantKey:     "tpu-v6e-slice",
-		},
-		{
-			machineType: "ct5p-hightpu-4t",
-			wantKey:     "tpu-v5p-slice",
-		},
-		{
-			machineType: "v5litepod-16",
-			wantKey:     "tpu-v5e-slice",
-		},
-		{
-			machineType: "n2-standard-4",
-			wantKey:     "google.com/tpu",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.machineType, func(t *testing.T) {
-			got := GetTPULimitKey(tt.machineType)
-			if got != tt.wantKey {
-				t.Errorf("GetTPULimitKey(%q) = %q, want %q", tt.machineType, got, tt.wantKey)
-			}
-		})
-	}
-}
-
-func TestGetGPULimitKey(t *testing.T) {
-	tests := []struct {
-		name        string
-		machineType string
-		accelLabel  string
-		wantKey     string
-		wantErr     bool
-	}{
-		{
-			name:        "a3-highgpu-8g",
-			machineType: "a3-highgpu-8g",
-			accelLabel:  "",
-			wantKey:     "nvidia-h100-80gb",
-			wantErr:     false,
-		},
-		{
-			name:        "a3-megagpu-8g",
-			machineType: "a3-megagpu-8g",
-			accelLabel:  "",
-			wantKey:     "nvidia-h100-mega-80gb",
-			wantErr:     false,
-		},
-		{
-			name:        "l4 fallback",
-			machineType: "custom-g2",
-			accelLabel:  "nvidia-l4",
-			wantKey:     "nvidia-l4",
-			wantErr:     false,
-		},
-		{
-			name:        "h100 fallback",
-			machineType: "custom-a3",
-			accelLabel:  "nvidia-h100-80gb",
-			wantKey:     "nvidia-h100-80gb",
-			wantErr:     false,
-		},
-		{
-			name:        "h200 fallback",
-			machineType: "custom-h200",
-			accelLabel:  "nvidia-h200-141gb",
-			wantKey:     "nvidia-h200-141gb",
-			wantErr:     false,
-		},
-		{
-			name:        "b200 fallback",
-			machineType: "custom-b200",
-			accelLabel:  "nvidia-b200",
-			wantKey:     "nvidia-b200",
-			wantErr:     false,
-		},
-		{
-			name:        "generic GPU fallback",
-			machineType: "custom-gpu",
-			accelLabel:  "",
-			wantKey:     "nvidia.com/gpu",
-			wantErr:     false,
-		},
-		{
-			name:        "reject TPU machine type",
-			machineType: "ct6e-standard-4t",
-			accelLabel:  "",
-			wantKey:     "",
-			wantErr:     true,
-		},
-		{
-			name:        "reject CPU n2-standard-4",
-			machineType: "n2-standard-4",
-			accelLabel:  "",
-			wantKey:     "",
-			wantErr:     true,
-		},
-		{
-			name:        "reject CPU e2-medium",
-			machineType: "e2-medium",
-			accelLabel:  "",
-			wantKey:     "",
-			wantErr:     true,
-		},
-		{
-			name:        "reject CPU c4-standard-4",
-			machineType: "c4-standard-4",
-			accelLabel:  "",
-			wantKey:     "",
-			wantErr:     true,
-		},
-		{
-			name:        "reject CPU n4-standard-4",
-			machineType: "n4-standard-4",
-			accelLabel:  "",
-			wantKey:     "",
-			wantErr:     true,
-		},
-		{
-			name:        "allow CPU c4-standard-4 with accelerator",
-			machineType: "c4-standard-4",
-			accelLabel:  "nvidia-l4",
-			wantKey:     "nvidia-l4",
-			wantErr:     false,
-		},
-		{
-			name:        "allow CPU n2-standard-4 with accelerator",
-			machineType: "n2-standard-4",
-			accelLabel:  "nvidia-l4",
-			wantKey:     "nvidia-l4",
-			wantErr:     false,
-		},
-		{
-			name:        "t4 fallback",
-			machineType: "custom-g1",
-			accelLabel:  "nvidia-tesla-t4",
-			wantKey:     "nvidia-tesla-t4",
-			wantErr:     false,
-		},
-		{
-			name:        "v100 fallback",
-			machineType: "custom-v100",
-			accelLabel:  "nvidia-tesla-v100",
-			wantKey:     "nvidia-tesla-v100",
-			wantErr:     false,
-		},
-		{
-			name:        "unknown accelerator label",
-			machineType: "custom-machine",
-			accelLabel:  "unknown-gpu",
-			wantKey:     "",
-			wantErr:     true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := GetGPULimitKey(tt.machineType, tt.accelLabel)
-			if (err != nil) != tt.wantErr {
-				t.Fatalf("GetGPULimitKey() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			if got != tt.wantKey {
-				t.Errorf("GetGPULimitKey() got = %q, want %q", got, tt.wantKey)
-			}
-		})
-	}
-}
